Move linePatterns into its own file
